course_allocation/requirements.txt
Flask==2.0.3
Flask-RESTful==0.3.9
mysql-connector-python==8.0.28
Werkzeug==2.0.3